Slimmed down large test files
