+++ bitloops_cli/src/engine/strategy/manual_commit/checkpoint_views.rs
@@ -1,1030 +1,4 @@
-fn get_git_author_from_repo(repo_root: &Path) -> Result<(String, String)> {
-    let local_name = run_git(repo_root, &["config", "--get", "user.name"]).ok();
-    let local_email = run_git(repo_root, &["config", "--get", "user.email"]).ok();
-    let global_name = run_git(repo_root, &["config", "--global", "--get", "user.name"]).ok();
-    let global_email = run_git(repo_root, &["config", "--global", "--get", "user.email"]).ok();
-
-    let name = local_name
-        .or(global_name)
-        .filter(|s| !s.trim().is_empty())
-        .unwrap_or_else(|| "Unknown".to_string());
-    let email = local_email
-        .or(global_email)
-        .filter(|s| !s.trim().is_empty())
-        .unwrap_or_else(|| "unknown@local".to_string());
-    Ok((name, email))
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct CheckpointSessionRef {
-    #[serde(default)]
-    pub metadata: String,
-    #[serde(default)]
-    pub transcript: String,
-    #[serde(default)]
-    pub context: String,
-    #[serde(default)]
-    pub content_hash: String,
-    #[serde(default)]
-    pub prompt: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct CheckpointSummaryView {
-    #[serde(default)]
-    pub checkpoint_id: String,
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub cli_version: String,
-    #[serde(default)]
-    pub strategy: String,
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub branch: String,
-    #[serde(default)]
-    pub checkpoints_count: u32,
-    #[serde(default)]
-    pub files_touched: Vec<String>,
-    #[serde(default)]
-    pub sessions: Vec<CheckpointSessionRef>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub token_usage: Option<TokenUsageMetadata>,
-    #[serde(default, skip)]
-    pub session_count: usize,
-}
-
-/// List-row view of a committed checkpoint (session-derived fields included).
-///
-/// Returned only by `list_committed()`. For single-checkpoint root metadata,
-/// use `read_committed()`.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct CommittedInfo {
-    #[serde(default)]
-    pub checkpoint_id: String,
-    #[serde(default)]
-    pub strategy: String,
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub branch: String,
-    #[serde(default)]
-    pub checkpoints_count: u32,
-    #[serde(default)]
-    pub files_touched: Vec<String>,
-    #[serde(default)]
-    pub session_count: usize,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub token_usage: Option<TokenUsageMetadata>,
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub session_id: String,
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub agent: String,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub agents: Vec<String>,
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub first_prompt_preview: String,
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub created_at: String,
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub is_task: bool,
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub tool_use_id: String,
-}
-
-fn summary_session_count(summary: &CheckpointSummaryView) -> usize {
-    summary.sessions.len()
-}
-
-#[allow(dead_code)]
-fn to_committed_info(
-    repo_root: &Path,
-    read_ref: &str,
-    summary: &CheckpointSummaryView,
-) -> CommittedInfo {
-    let mut info = CommittedInfo {
-        checkpoint_id: summary.checkpoint_id.clone(),
-        strategy: summary.strategy.clone(),
-        branch: summary.branch.clone(),
-        checkpoints_count: summary.checkpoints_count,
-        files_touched: summary.files_touched.clone(),
-        session_count: summary_session_count(summary),
-        token_usage: summary.token_usage.clone(),
-        ..Default::default()
-    };
-
-    if info.session_count == 0 {
-        return info;
-    }
-
-    let (a, b) = checkpoint_dir_parts(&summary.checkpoint_id);
-    let latest_session_index = info.session_count - 1;
-
-    for idx in 0..info.session_count {
-        let meta_path = format!("{a}/{b}/{idx}/{}", paths::METADATA_FILE_NAME);
-        let Ok(raw) = git_show_file(repo_root, read_ref, &meta_path) else {
-            continue;
-        };
-
-        if let Ok(meta) = serde_json::from_str::<CommittedMetadata>(&raw) {
-            push_unique_agent(&mut info.agents, &meta.agent);
-            if idx == latest_session_index {
-                info.session_id = meta.session_id;
-                info.agent = canonicalize_agent_type(&meta.agent);
-                info.created_at = meta.created_at;
-                info.is_task = meta.is_task;
-                info.tool_use_id = meta.tool_use_id;
-            }
-            continue;
-        }
-
-        // Keep list/read behavior resilient to legacy metadata with partial fields.
-        if let Ok(meta) = serde_json::from_str::<serde_json::Value>(&raw) {
-            push_unique_agent(
-                &mut info.agents,
-                meta.get("agent")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or_default(),
-            );
-
-            if idx == latest_session_index {
-                info.session_id = meta
-                    .get("session_id")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or_default()
-                    .to_string();
-                info.agent = canonicalize_agent_type(
-                    meta.get("agent")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or_default(),
-                );
-                info.created_at = meta
-                    .get("created_at")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or_default()
-                    .to_string();
-                info.is_task = meta
-                    .get("is_task")
-                    .and_then(serde_json::Value::as_bool)
-                    .unwrap_or(false);
-                info.tool_use_id = meta
-                    .get("tool_use_id")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or_default()
-                    .to_string();
-            }
-        }
-    }
-
-    if info.agent.is_empty()
-        && let Some(last) = info.agents.last()
-    {
-        info.agent = last.clone();
-    }
-
-    let first_prompt_path = format!("{a}/{b}/0/{}", paths::PROMPT_FILE_NAME);
-    if let Ok(raw_prompts) = git_show_file(repo_root, read_ref, &first_prompt_path) {
-        info.first_prompt_preview = first_prompt_preview(&raw_prompts);
-    }
-
-    info
-}
-
-fn push_unique_agent(agents: &mut Vec<String>, agent: &str) {
-    let normalized = canonicalize_agent_type(agent);
-    if normalized.is_empty() || agents.iter().any(|existing| existing == &normalized) {
-        return;
-    }
-    agents.push(normalized);
-}
-
-fn first_prompt_preview(prompts_blob: &str) -> String {
-    let first_prompt = prompts_blob.split("\n\n---\n\n").next().unwrap_or_default();
-    let stripped = strip_leading_wrapped_tags(first_prompt).trim_start();
-    stripped.chars().take(160).collect()
-}
-
-fn strip_leading_wrapped_tags(input: &str) -> &str {
-    let mut rest = input.trim_start();
-    loop {
-        let Some((tag_name, after_open)) = parse_leading_open_tag(rest) else {
-            return rest;
-        };
-        let closing_tag = format!("</{tag_name}>");
-        let Some(close_idx) = after_open.find(&closing_tag) else {
-            return rest;
-        };
-        rest = after_open[close_idx + closing_tag.len()..].trim_start();
-    }
-}
-
-fn parse_leading_open_tag(input: &str) -> Option<(String, &str)> {
-    let trimmed = input.trim_start();
-    if !trimmed.starts_with('<') || trimmed.starts_with("</") {
-        return None;
-    }
-
-    let open_end = trimmed.find('>')?;
-    let inside = trimmed.get(1..open_end)?.trim();
-    if inside.is_empty() || inside.ends_with('/') {
-        return None;
-    }
-
-    let tag_name = inside.split_whitespace().next()?.trim();
-    if tag_name.is_empty() {
-        return None;
-    }
-
-    Some((tag_name.to_string(), trimmed.get(open_end + 1..)?))
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct CheckpointAuthor {
-    pub name: String,
-    pub email: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct SessionContentView {
-    pub metadata: serde_json::Value,
-    pub transcript: String,
-    pub prompts: String,
-    pub context: String,
-}
-
-pub fn update_summary(
-    repo_root: &Path,
-    checkpoint_id: &str,
-    summary: serde_json::Value,
-) -> Result<()> {
-    let redacted_summary = redact_json_value(&summary);
-    let db_updated =
-        update_checkpoint_session_summary_in_db(repo_root, checkpoint_id, &redacted_summary)?;
-    if db_updated {
-        return Ok(());
-    }
-    anyhow::bail!("checkpoint not found: {checkpoint_id}");
-}
-
-fn build_checkpoint_session_ref(checkpoint_id: &str, session_index: i64) -> CheckpointSessionRef {
-    let (a, b) = checkpoint_dir_parts(checkpoint_id);
-    let base = format!("{a}/{b}/{session_index}");
-    CheckpointSessionRef {
-        metadata: format!("/{base}/{}", paths::METADATA_FILE_NAME),
-        transcript: format!("/{base}/{}", paths::TRANSCRIPT_FILE_NAME),
-        context: format!("/{base}/{}", paths::CONTEXT_FILE_NAME),
-        content_hash: format!("/{base}/{}", paths::CONTENT_HASH_FILE_NAME),
-        prompt: format!("/{base}/{}", paths::PROMPT_FILE_NAME),
-    }
-}
-
-fn parse_json_string_array(raw: &str) -> Vec<String> {
-    serde_json::from_str::<Vec<String>>(raw).unwrap_or_default()
-}
-
-fn read_checkpoint_blob_text(
-    storage: &CheckpointStorageContext,
-    checkpoint_id: &str,
-    session_index: i64,
-    blob_type: crate::engine::blob::BlobType,
-) -> String {
-    let reference = crate::engine::blob::load_checkpoint_blob_reference(
-        &storage.sqlite,
-        checkpoint_id,
-        session_index,
-        blob_type.as_str(),
-    );
-    let Ok(Some(reference)) = reference else {
-        return String::new();
-    };
-    let Ok(bytes) = storage.blob_store.read(&reference.storage_path) else {
-        return String::new();
-    };
-    String::from_utf8_lossy(&bytes).to_string()
-}
-
-fn read_committed_from_db(
-    storage: &CheckpointStorageContext,
-    checkpoint_id: &str,
-) -> Result<Option<CheckpointSummaryView>> {
-    use rusqlite::OptionalExtension;
-
-    let checkpoint_row = storage.sqlite.with_connection(|conn| {
-        conn.query_row(
-            "SELECT strategy, branch, cli_version, checkpoints_count, files_touched, token_usage
-             FROM checkpoints
-             WHERE checkpoint_id = ?1 AND repo_id = ?2
-             LIMIT 1",
-            rusqlite::params![checkpoint_id, storage.repo_id.as_str()],
-            |row| {
-                Ok((
-                    row.get::<_, String>(0)?,
-                    row.get::<_, String>(1)?,
-                    row.get::<_, String>(2)?,
-                    row.get::<_, i64>(3)?,
-                    row.get::<_, String>(4)?,
-                    row.get::<_, Option<String>>(5)?,
-                ))
-            },
-        )
-        .optional()
-        .map_err(anyhow::Error::from)
-    })?;
-    let Some((strategy, branch, cli_version, checkpoints_count, files_touched_raw, token_usage_raw)) =
-        checkpoint_row
-    else {
-        return Ok(None);
-    };
-
-    let session_indexes = storage.sqlite.with_connection(|conn| {
-        let mut stmt = conn.prepare(
-            "SELECT session_index
-             FROM checkpoint_sessions
-             WHERE checkpoint_id = ?1
-             ORDER BY session_index ASC",
-        )?;
-        let mut rows = stmt.query(rusqlite::params![checkpoint_id])?;
-        let mut indexes: Vec<i64> = Vec::new();
-        while let Some(row) = rows.next()? {
-            indexes.push(row.get::<_, i64>(0)?);
-        }
-        Ok(indexes)
-    })?;
-
-    let sessions = session_indexes
-        .into_iter()
-        .map(|idx| build_checkpoint_session_ref(checkpoint_id, idx))
-        .collect::<Vec<_>>();
-    let token_usage = token_usage_raw
-        .as_deref()
-        .and_then(|raw| serde_json::from_str::<TokenUsageMetadata>(raw).ok());
-    let mut summary = CheckpointSummaryView {
-        checkpoint_id: checkpoint_id.to_string(),
-        cli_version,
-        strategy,
-        branch,
-        checkpoints_count: checkpoints_count.max(0).min(u32::MAX as i64) as u32,
-        files_touched: parse_json_string_array(&files_touched_raw),
-        sessions,
-        token_usage,
-        ..Default::default()
-    };
-    summary.session_count = summary.sessions.len();
-    Ok(Some(summary))
-}
-
-fn to_committed_info_from_db(
-    storage: &CheckpointStorageContext,
-    summary: &CheckpointSummaryView,
-) -> Result<CommittedInfo> {
-    let mut info = CommittedInfo {
-        checkpoint_id: summary.checkpoint_id.clone(),
-        strategy: summary.strategy.clone(),
-        branch: summary.branch.clone(),
-        checkpoints_count: summary.checkpoints_count,
-        files_touched: summary.files_touched.clone(),
-        session_count: summary.sessions.len(),
-        token_usage: summary.token_usage.clone(),
-        ..Default::default()
-    };
-    if info.session_count == 0 {
-        return Ok(info);
-    }
-
-    let session_rows = storage.sqlite.with_connection(|conn| {
-        let mut stmt = conn.prepare(
-            "SELECT session_id, agent, created_at, is_task, tool_use_id, session_index
-             FROM checkpoint_sessions
-             WHERE checkpoint_id = ?1
-             ORDER BY session_index ASC",
-        )?;
-        let mut rows = stmt.query(rusqlite::params![summary.checkpoint_id])?;
-        let mut sessions: Vec<(String, String, String, bool, String, i64)> = Vec::new();
-        while let Some(row) = rows.next()? {
-            let is_task = row.get::<_, i64>(3)? != 0;
-            sessions.push((
-                row.get::<_, String>(0)?,
-                row.get::<_, String>(1)?,
-                row.get::<_, String>(2)?,
-                is_task,
-                row.get::<_, String>(4)?,
-                row.get::<_, i64>(5)?,
-            ));
-        }
-        Ok(sessions)
-    })?;
-
-    for (_, agent, _, _, _, _) in &session_rows {
-        push_unique_agent(&mut info.agents, agent);
-    }
-    if let Some((session_id, agent, created_at, is_task, tool_use_id, _)) = session_rows.last() {
-        info.session_id = session_id.clone();
-        info.agent = canonicalize_agent_type(agent);
-        info.created_at = created_at.clone();
-        info.is_task = *is_task;
-        info.tool_use_id = tool_use_id.clone();
-    }
-    if info.agent.is_empty()
-        && let Some(last) = info.agents.last()
-    {
-        info.agent = last.clone();
-    }
-
-    if let Some(first_ref) = crate::engine::blob::load_checkpoint_blob_reference(
-        &storage.sqlite,
-        &summary.checkpoint_id,
-        0,
-        crate::engine::blob::BlobType::Prompts.as_str(),
-    )? && let Ok(prompt_bytes) = storage.blob_store.read(&first_ref.storage_path)
-    {
-        info.first_prompt_preview = first_prompt_preview(&String::from_utf8_lossy(&prompt_bytes));
-    }
-
-    Ok(info)
-}
-
-fn read_session_content_from_db(
-    storage: &CheckpointStorageContext,
-    checkpoint_id: &str,
-    session_index: usize,
-) -> Result<Option<SessionContentView>> {
-    use rusqlite::OptionalExtension;
-
-    let session_row = storage.sqlite.with_connection(|conn| {
-        conn.query_row(
-            "SELECT c.strategy, c.branch, c.cli_version,
-                    s.session_id, s.agent, s.created_at, s.turn_id, s.checkpoints_count,
-                    s.files_touched, s.is_task, s.tool_use_id,
-                    s.transcript_identifier_at_start, s.checkpoint_transcript_start,
-                    s.initial_attribution, s.token_usage, s.summary,
-                    s.transcript_path
-             FROM checkpoint_sessions s
-             JOIN checkpoints c ON c.checkpoint_id = s.checkpoint_id
-             WHERE s.checkpoint_id = ?1
-               AND s.session_index = ?2
-               AND c.repo_id = ?3
-             LIMIT 1",
-            rusqlite::params![checkpoint_id, session_index as i64, storage.repo_id.as_str()],
-            |row| {
-                Ok((
-                    row.get::<_, String>(0)?,
-                    row.get::<_, String>(1)?,
-                    row.get::<_, String>(2)?,
-                    row.get::<_, String>(3)?,
-                    row.get::<_, String>(4)?,
-                    row.get::<_, String>(5)?,
-                    row.get::<_, String>(6)?,
-                    row.get::<_, i64>(7)?,
-                    row.get::<_, String>(8)?,
-                    row.get::<_, i64>(9)?,
-                    row.get::<_, String>(10)?,
-                    row.get::<_, String>(11)?,
-                    row.get::<_, i64>(12)?,
-                    row.get::<_, Option<String>>(13)?,
-                    row.get::<_, Option<String>>(14)?,
-                    row.get::<_, Option<String>>(15)?,
-                    row.get::<_, String>(16)?,
-                ))
-            },
-        )
-        .optional()
-        .map_err(anyhow::Error::from)
-    })?;
-    let Some((
-        strategy,
-        branch,
-        cli_version,
-        session_id,
-        agent,
-        created_at,
-        turn_id,
-        checkpoints_count,
-        files_touched_raw,
-        is_task,
-        tool_use_id,
-        transcript_identifier_at_start,
-        checkpoint_transcript_start,
-        initial_attribution_raw,
-        token_usage_raw,
-        summary_raw,
-        transcript_path,
-    )) = session_row
-    else {
-        return Ok(None);
-    };
-
-    let metadata = CommittedMetadata {
-        checkpoint_id: checkpoint_id.to_string(),
-        session_id,
-        checkpoints_count: checkpoints_count.max(0).min(u32::MAX as i64) as u32,
-        strategy,
-        agent,
-        created_at,
-        cli_version,
-        turn_id,
-        files_touched: parse_json_string_array(&files_touched_raw),
-        is_task: is_task != 0,
-        tool_use_id,
-        transcript_identifier_at_start,
-        checkpoint_transcript_start,
-        transcript_lines_at_start: checkpoint_transcript_start,
-        branch,
-        summary: summary_raw
-            .as_deref()
-            .and_then(|raw| serde_json::from_str::<serde_json::Value>(raw).ok()),
-        token_usage: token_usage_raw
-            .as_deref()
-            .and_then(|raw| serde_json::from_str::<TokenUsageMetadata>(raw).ok()),
-        initial_attribution: initial_attribution_raw
-            .as_deref()
-            .and_then(|raw| serde_json::from_str::<serde_json::Value>(raw).ok()),
-        transcript_path,
-    };
-    let metadata_value =
-        serde_json::to_value(&metadata).context("serializing checkpoint session metadata from DB")?;
-
-    let transcript = read_checkpoint_blob_text(
-        storage,
-        checkpoint_id,
-        session_index as i64,
-        crate::engine::blob::BlobType::Transcript,
-    );
-    let prompts = read_checkpoint_blob_text(
-        storage,
-        checkpoint_id,
-        session_index as i64,
-        crate::engine::blob::BlobType::Prompts,
-    );
-    let context = read_checkpoint_blob_text(
-        storage,
-        checkpoint_id,
-        session_index as i64,
-        crate::engine::blob::BlobType::Context,
-    );
-
-    Ok(Some(SessionContentView {
-        metadata: metadata_value,
-        transcript,
-        prompts,
-        context,
-    }))
-}
-
-pub fn list_committed(repo_root: &Path) -> Result<Vec<CommittedInfo>> {
-    let storage = open_checkpoint_storage_context(repo_root)?;
-    let checkpoint_ids = storage.sqlite.with_connection(|conn| {
-        let mut stmt = conn.prepare(
-            "SELECT checkpoint_id
-             FROM checkpoints
-             WHERE repo_id = ?1
-             ORDER BY created_at DESC, checkpoint_id DESC",
-        )?;
-        let mut rows = stmt.query(rusqlite::params![storage.repo_id.as_str()])?;
-        let mut ids = Vec::new();
-        while let Some(row) = rows.next()? {
-            ids.push(row.get::<_, String>(0)?);
-        }
-        Ok(ids)
-    })?;
-
-    let mut out: Vec<CommittedInfo> = Vec::new();
-    for checkpoint_id in checkpoint_ids {
-        if let Some(summary) = read_committed_from_db(&storage, &checkpoint_id)? {
-            out.push(to_committed_info_from_db(&storage, &summary)?);
-        }
-    }
-    Ok(out)
-}
-
-pub fn get_checkpoint_author(repo_root: &Path, checkpoint_id: &str) -> Result<CheckpointAuthor> {
-    use rusqlite::OptionalExtension;
-
-    let storage = open_checkpoint_storage_context(repo_root)?;
-    let db_author = storage.sqlite.with_connection(|conn| {
-        conn.query_row(
-            "SELECT s.author_name, s.author_email
-             FROM checkpoint_sessions s
-             JOIN checkpoints c ON c.checkpoint_id = s.checkpoint_id
-             WHERE s.checkpoint_id = ?1
-               AND c.repo_id = ?2
-             ORDER BY s.session_index ASC
-             LIMIT 1",
-            rusqlite::params![checkpoint_id, storage.repo_id.as_str()],
-            |row| Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?)),
-        )
-        .optional()
-        .map_err(anyhow::Error::from)
-    })?;
-    if let Some((name, email)) = db_author
-        && (!name.trim().is_empty() || !email.trim().is_empty())
-    {
-        return Ok(CheckpointAuthor { name, email });
-    }
-    Ok(CheckpointAuthor::default())
-}
-
-pub fn read_committed(
-    repo_root: &Path,
-    checkpoint_id: &str,
-) -> Result<Option<CheckpointSummaryView>> {
-    let storage = open_checkpoint_storage_context(repo_root)?;
-    read_committed_from_db(&storage, checkpoint_id)
-}
-
-#[allow(dead_code)]
-fn read_committed_with_ref(
-    repo_root: &Path,
-    read_ref: &str,
-    checkpoint_id: &str,
-) -> Result<Option<CheckpointSummaryView>> {
-    if checkpoint_type_for_ref(read_ref) != CheckpointType::Committed {
-        return Ok(None);
-    }
-
-    let (a, b) = checkpoint_dir_parts(checkpoint_id);
-    let metadata_path = format!("{a}/{b}/{}", paths::METADATA_FILE_NAME);
-    let raw = match git_show_file(repo_root, read_ref, &metadata_path) {
-        Ok(v) => v,
-        Err(_) => return Ok(None),
-    };
-
-    let mut summary: CheckpointSummaryView = serde_json::from_str(&raw)
-        .with_context(|| format!("parsing checkpoint {checkpoint_id}"))?;
-    summary.session_count = summary.sessions.len();
-    Ok(Some(summary))
-}
-
-/// Returns one committed checkpoint in list shape (session-derived fields included).
-pub fn read_committed_info(repo_root: &Path, checkpoint_id: &str) -> Result<Option<CommittedInfo>> {
-    let storage = open_checkpoint_storage_context(repo_root)?;
-    if let Some(summary) = read_committed_from_db(&storage, checkpoint_id)? {
-        return Ok(Some(to_committed_info_from_db(&storage, &summary)?));
-    }
-    Ok(None)
-}
-
-pub fn read_session_content(
-    repo_root: &Path,
-    checkpoint_id: &str,
-    session_index: usize,
-) -> Result<SessionContentView> {
-    let storage = open_checkpoint_storage_context(repo_root)?;
-    let summary = read_committed_from_db(&storage, checkpoint_id)?
-        .ok_or_else(|| anyhow::anyhow!("checkpoint not found"))?;
-    let session_count = summary_session_count(&summary);
-    if session_index >= session_count {
-        anyhow::bail!("session {session_index} not found");
-    }
-    read_session_content_from_db(&storage, checkpoint_id, session_index)?
-        .ok_or_else(|| anyhow::anyhow!("session {session_index} not found"))
-}
-
-pub fn read_latest_session_content(
-    repo_root: &Path,
-    checkpoint_id: &str,
-) -> Result<SessionContentView> {
-    let summary = read_committed(repo_root, checkpoint_id)?
-        .ok_or_else(|| anyhow::anyhow!("checkpoint not found"))?;
-    let session_count = summary_session_count(&summary);
-    if session_count == 0 {
-        anyhow::bail!("checkpoint has no sessions");
-    }
-    read_session_content(repo_root, checkpoint_id, session_count - 1)
-}
-
-pub fn read_session_content_by_id(
-    repo_root: &Path,
-    checkpoint_id: &str,
-    session_id: &str,
-) -> Result<SessionContentView> {
-    let summary = read_committed(repo_root, checkpoint_id)?
-        .ok_or_else(|| anyhow::anyhow!("checkpoint not found"))?;
-    let session_count = summary_session_count(&summary);
-    for idx in 0..session_count {
-        // Skip unreadable session slots while searching by session ID.
-        let Ok(content) = read_session_content(repo_root, checkpoint_id, idx) else {
-            continue;
-        };
-        if content
-            .metadata
-            .get("session_id")
-            .and_then(serde_json::Value::as_str)
-            == Some(session_id)
-        {
-            return Ok(content);
-        }
-    }
-    anyhow::bail!("session {session_id:?} not found in checkpoint {checkpoint_id}")
-}
-
-pub fn read_commit_checkpoint_mappings(repo_root: &Path) -> Result<std::collections::HashMap<String, String>> {
-    let sqlite_path = resolve_temporary_checkpoint_sqlite_path(repo_root)?;
-    let sqlite = crate::engine::db::SqliteConnectionPool::connect(sqlite_path)
-        .context("opening SQLite database for commit-checkpoint mappings")?;
-    sqlite
-        .initialise_checkpoint_schema()
-        .context("initialising checkpoint schema for commit-checkpoint mappings")?;
-
-    let repo_id = crate::engine::devql::resolve_repo_id(repo_root)
-        .context("resolving repo identity for commit-checkpoint mappings")?;
-
-    sqlite.with_connection(|conn| {
-        let mut stmt = conn.prepare(
-            "SELECT commit_sha, checkpoint_id
-             FROM commit_checkpoints
-             WHERE repo_id = ?1
-             ORDER BY created_at DESC, checkpoint_id DESC",
-        )?;
-        let mut rows = stmt.query(rusqlite::params![repo_id.as_str()])?;
-
-        let mut out: std::collections::HashMap<String, String> = std::collections::HashMap::new();
-        while let Some(row) = rows.next()? {
-            let commit_sha = row.get::<_, String>(0)?.trim().to_string();
-            let checkpoint_id = row.get::<_, String>(1)?.trim().to_string();
-            if commit_sha.is_empty() || !is_valid_checkpoint_id(&checkpoint_id) {
-                continue;
-            }
-            out.entry(commit_sha).or_insert(checkpoint_id);
-        }
-        Ok(out)
-    })
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-struct CodeLearning {
-    path: String,
-    #[serde(default)]
-    line: u32,
-    #[serde(default)]
-    end_line: u32,
-    finding: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-struct LearningsSummary {
-    repo: Vec<String>,
-    code: Vec<CodeLearning>,
-    workflow: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-struct Summary {
-    intent: String,
-    outcome: String,
-    learnings: LearningsSummary,
-    friction: Vec<String>,
-    open_items: Vec<String>,
-}
-
-fn redact_summary(summary: Option<&Summary>) -> Result<Option<Summary>> {
-    let Some(summary) = summary else {
-        return Ok(None);
-    };
-    Ok(Some(Summary {
-        intent: redact_text(&summary.intent),
-        outcome: redact_text(&summary.outcome),
-        learnings: LearningsSummary {
-            repo: redact_string_slice(Some(&summary.learnings.repo))?.unwrap_or_default(),
-            code: redact_code_learnings(Some(&summary.learnings.code))?.unwrap_or_default(),
-            workflow: redact_string_slice(Some(&summary.learnings.workflow))?.unwrap_or_default(),
-        },
-        friction: redact_string_slice(Some(&summary.friction))?.unwrap_or_default(),
-        open_items: redact_string_slice(Some(&summary.open_items))?.unwrap_or_default(),
-    }))
-}
-
-fn redact_string_slice(values: Option<&[String]>) -> Result<Option<Vec<String>>> {
-    let Some(values) = values else {
-        return Ok(None);
-    };
-    Ok(Some(
-        values.iter().map(|value| redact_text(value)).collect(),
-    ))
-}
-
-fn redact_code_learnings(values: Option<&[CodeLearning]>) -> Result<Option<Vec<CodeLearning>>> {
-    let Some(values) = values else {
-        return Ok(None);
-    };
-    Ok(Some(
-        values
-            .iter()
-            .map(|value| CodeLearning {
-                path: value.path.clone(),
-                line: value.line,
-                end_line: value.end_line,
-                finding: redact_text(&value.finding),
-            })
-            .collect(),
-    ))
-}
-
-fn copy_metadata_dir(
-    metadata_dir: &Path,
-    base_path: &str,
-) -> Result<std::collections::BTreeMap<String, String>> {
-    add_directory_to_entries_with_abs_path(metadata_dir, base_path)
-}
-
-fn add_directory_to_entries_with_abs_path(
-    metadata_dir: &Path,
-    base_path: &str,
-) -> Result<std::collections::BTreeMap<String, String>> {
-    let mut out: std::collections::BTreeMap<String, String> = std::collections::BTreeMap::new();
-    if !metadata_dir.exists() {
-        return Ok(out);
-    }
-
-    let mut stack: Vec<PathBuf> = vec![metadata_dir.to_path_buf()];
-    while let Some(dir) = stack.pop() {
-        for entry in fs::read_dir(&dir).with_context(|| format!("reading {}", dir.display()))? {
-            let entry = entry?;
-            let path = entry.path();
-            let lmeta = fs::symlink_metadata(&path)?;
-            if lmeta.file_type().is_symlink() {
-                continue;
-            }
-            if lmeta.is_dir() {
-                stack.push(path);
-                continue;
-            }
-
-            let rel = path
-                .strip_prefix(metadata_dir)
-                .with_context(|| format!("path traversal detected: {}", path.display()))?;
-            let rel = rel.to_string_lossy().replace('\\', "/");
-            if rel.starts_with("..") {
-                anyhow::bail!("path traversal detected: {rel}");
-            }
-            let key = format!(
-                "{}/{}",
-                base_path.trim_end_matches('/'),
-                rel.trim_start_matches('/')
-            );
-            let content = fs::read(&path)?;
-            let redacted_bytes = if key.ends_with(".jsonl") {
-                redact_jsonl_bytes_with_fallback(&content)
-            } else {
-                redact_bytes(&content)
-            };
-            let redacted = String::from_utf8_lossy(&redacted_bytes).to_string();
-            out.insert(key, redacted);
-        }
-    }
-
-    Ok(out)
-}
-
-const EMPTY_TREE_HASH: &str = "4b825dc642cb6eb9a060e54bf8d69288fbee4904";
-
-fn ensure_metadata_branch(repo_root: &Path) -> Result<()> {
-    let metadata_ref = format!("refs/heads/{}", paths::METADATA_BRANCH_NAME);
-    if run_git(repo_root, &["rev-parse", &metadata_ref]).is_ok() {
-        return Ok(());
-    }
-    let (author_name, author_email) = get_git_author_from_repo(repo_root)?;
-    let commit = run_git_env(
-        repo_root,
-        &[
-            "commit-tree",
-            EMPTY_TREE_HASH,
-            "-m",
-            "Initialize checkpoints branch",
-        ],
-        &[
-            ("GIT_AUTHOR_NAME", &author_name),
-            ("GIT_AUTHOR_EMAIL", &author_email),
-            ("GIT_COMMITTER_NAME", &author_name),
-            ("GIT_COMMITTER_EMAIL", &author_email),
-        ],
-    )?;
-    run_git(repo_root, &["update-ref", &metadata_ref, commit.trim()])?;
-    Ok(())
-}
-
-pub(crate) fn commit_files_to_metadata_branch(
-    repo_root: &Path,
-    files: &[(PathBuf, String)],
-    commit_message: &str,
-    author_name: &str,
-    author_email: &str,
-) -> Result<()> {
-    ensure_metadata_branch(repo_root)?;
-    let metadata_ref = format!("refs/heads/{}", paths::METADATA_BRANCH_NAME);
-    let parent_tree = run_git(
-        repo_root,
-        &["rev-parse", &format!("{metadata_ref}^{{tree}}")],
-    )
-    .ok()
-    .filter(|s| !s.is_empty());
-    let parent_commit = run_git(repo_root, &["rev-parse", &metadata_ref])
-        .ok()
-        .filter(|s| !s.is_empty());
-
-    let tree = build_tree_with_explicit_paths(repo_root, parent_tree.as_deref(), files)?;
-
-    let mut ct_args: Vec<String> = vec!["commit-tree".into(), tree];
-    if let Some(parent) = parent_commit {
-        ct_args.push("-p".into());
-        ct_args.push(parent);
-    }
-    ct_args.push("-m".into());
-    ct_args.push(commit_message.to_string());
-    let ct_args_ref: Vec<&str> = ct_args.iter().map(String::as_str).collect();
-    let commit = run_git_env(
-        repo_root,
-        &ct_args_ref,
-        &[
-            ("GIT_AUTHOR_NAME", author_name),
-            ("GIT_AUTHOR_EMAIL", author_email),
-            ("GIT_COMMITTER_NAME", author_name),
-            ("GIT_COMMITTER_EMAIL", author_email),
-        ],
-    )?;
-    run_git(repo_root, &["update-ref", &metadata_ref, commit.trim()])?;
-    Ok(())
-}
-
-#[allow(dead_code)]
-pub(crate) fn git_show_file(repo_root: &Path, reference: &str, tree_path: &str) -> Result<String> {
-    run_git(repo_root, &["show", &format!("{reference}:{tree_path}")])
-}
-
-pub(crate) fn git_show_file_bytes(
-    repo_root: &Path,
-    reference: &str,
-    tree_path: &str,
-) -> Result<Vec<u8>> {
-    let output = new_git_command()
-        .args(["show", &format!("{reference}:{tree_path}")])
-        .current_dir(repo_root)
-        .stdin(Stdio::null())
-        .output()
-        .with_context(|| format!("running git show {reference}:{tree_path}"))?;
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        anyhow::bail!(
-            "git show {reference}:{tree_path} failed ({}): {}",
-            output.status,
-            stderr.trim()
-        );
-    }
-    Ok(output.stdout)
-}
-
-fn get_commit_author(repo_root: &Path, commit_ref: &str) -> Option<(String, String)> {
-    let raw = run_git(repo_root, &["show", "-s", "--format=%an%n%ae", commit_ref]).ok()?;
-    let mut lines = raw.lines();
-    let name = lines.next().unwrap_or_default().trim().to_string();
-    let email = lines.next().unwrap_or_default().trim().to_string();
-    if name.is_empty() || email.is_empty() {
-        return None;
-    }
-    Some((name, email))
-}
-
-#[allow(dead_code)]
-pub(crate) fn metadata_read_ref(repo_root: &Path) -> Option<String> {
-    if !crate::engine::session::legacy_local_backend_enabled() {
-        return None;
-    }
-
-    let local = format!("refs/heads/{}", paths::METADATA_BRANCH_NAME);
-    if run_git(repo_root, &["rev-parse", &local]).is_ok() {
-        return Some(local);
-    }
-    let remote = format!("refs/remotes/origin/{}", paths::METADATA_BRANCH_NAME);
-    if run_git(repo_root, &["rev-parse", &remote]).is_ok() {
-        return Some(remote);
-    }
-    None
-}
-
-fn current_branch_name(repo_root: &Path) -> String {
-    run_git(repo_root, &["symbolic-ref", "--quiet", "--short", "HEAD"]).unwrap_or_default()
-}
-
-fn redact_json_value(value: &serde_json::Value) -> serde_json::Value {
-    match value {
-        serde_json::Value::String(s) => serde_json::Value::String(redact_text(s)),
-        serde_json::Value::Array(arr) => {
-            serde_json::Value::Array(arr.iter().map(redact_json_value).collect())
-        }
-        serde_json::Value::Object(map) => serde_json::Value::Object(
-            map.iter()
-                .map(|(k, v)| (k.clone(), redact_json_value(v)))
-                .collect(),
-        ),
-        _ => value.clone(),
-    }
-}
-
-fn redact_bytes(input: &[u8]) -> Vec<u8> {
-    redact::bytes(input).into_owned()
-}
-
-fn redact_jsonl_bytes_with_fallback(input: &[u8]) -> Vec<u8> {
-    match redact::jsonl_bytes(input) {
-        Ok(redacted) => redacted.into_owned(),
-        Err(_) => redact::bytes(input).into_owned(),
-    }
-}
-
-fn redact_text(input: &str) -> String {
-    redact::string(input)
-}
+include!("checkpoint_views/types.rs");
+include!("checkpoint_views/readers.rs");
+include!("checkpoint_views/mappings.rs");
+include!("checkpoint_views/metadata_branch.rs");
